knowledge-qa/requirements.txt
langchain>=0.3.0
langchain-core>=0.3.0
langchain-google-genai>=2.0.0
langchain-community>=0.3.0
langchain-chroma>=0.1.0
langchain-text-splitters>=0.3.0
chromadb>=0.4.0
python-dotenv>=1.0.0
rich>=13.0.0
pypdf>=3.0.0
tiktoken>=0.5.0
